MPI enabled code

+++ utils/dftParameters.cc
@@ -541,7 +541,7 @@ namespace dftfe
           prm.declare_entry(
             "ATOM BALL RADIUS",
             "0.0",
-            Patterns::Double(0, 20),
+            Patterns::Double(0, 100),
             "[Standard] Radius of ball enclosing every atom, inside which the mesh size is set close to MESH SIZE AROUND ATOM and coarse-grained in the region outside the enclosing balls. For the default value of 0.0, a heuristically determined value is used, which is good enough for most cases but can be a bit conservative choice for fully non-periodic and semi-periodic problems as well as all-electron problems. To improve the computational efficiency user may experiment with values of ATOM BALL RADIUS ranging between 3.0 to 6.0 for pseudopotential problems, and ranging between 1.0 to 2.5 for all-electron problems.  Units: a.u.");
 
           prm.declare_entry(
